requirements.txt
websockets
Pillow
